WIP: major refactor of Map.js, splitting it up into multiple components, adding eslint/prettier, adding utils for the map and esri/airtable integrations
