+++ eslint.config.js
@@ -1,9 +1,12 @@
+import globals from "globals";
 import eslintConfig from "@mapc/eslint-config";
 
 export default [
+  ...eslintConfig,
   {
-    plugins: {
-      "mapc-eslint-config": eslintConfig,
+    languageOptions: { globals: globals.node },
+    rules: {
+      "react/prop-types": "off",
     },
   },
 ];
